refactor: get{}BasicInfo를 dto에서 helper로 옮김
