--- server/apps/api/src/channel/dto/channel-basic-info.dto.ts
@@ -1,11 +1,0 @@
-export const getChannelBasicInfo = (channel) => {
-  return {
-    _id: channel._id,
-    name: channel.name,
-    managerId: channel.managerId,
-    type: channel.type,
-    isPrivate: channel.isPrivate,
-    profileUrl: channel.profileUrl,
-    description: channel.description,
-  };
-};
