+++ server/apps/api/src/channel/helper/index.ts
@@ -1,0 +1,2 @@
+export * from './addObjectForm';
+export * from './getChannelBasicInfo';
