+++ server/apps/api/src/community/community.controller.ts
@@ -2,6 +2,7 @@ import {
   Body,
   Controller,
   Delete,
+  Get,
   Inject,
   LoggerService,
   Param,
