+++ server/apps/api/src/user/helper/getUserBasicInfo.ts
@@ -1,0 +1,10 @@
+export const getUserBasicInfo = (user) => {
+  return {
+    _id: user._id,
+    id: user.id,
+    nickname: user.nickname,
+    status: user.status,
+    profileUrl: user.profileUrl,
+    description: user.description,
+  };
+};
